requirements.txt
eth-brownie
python-dotenv